# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "SGL API"
    },
    "paths": {
...
        "/lists/{listID}/items": {
            "get": {
                "tags": [
                    "Items"
                ],
                "description": "Get all items in a list",
                "parameters": [
                    {
                        "name": "listID",
                        "in": "path",
                        "description": "id of list",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "description": "Create an item.",
                "parameters": [
                    {
                        "name": "listID",
                        "in": "path",
                        "description": "id of list",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "name of list",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "Items"
                ]
            }
        }
    }
}